Decouple tvscreener retry and capability state
TvScreenerService mixed query execution with retry policy and capability-cache primitives. This keeps the public service API stable while moving retry exceptions/executor and capability state objects into focused modules for narrower future changes.\n\nConstraint: tvscreener query semantics and MCP screening fallbacks must remain unchanged.\nRejected: Split stock/crypto query methods now | retry and capability extraction gives a smaller first cut with direct test coverage.\nConfidence: high\nScope-risk: moderate\nReversibility: clean\nDirective: Do not alter tvscreener field aliases or fallback behavior without rerunning MCP screening contract tests.\nTested: uv run --group dev ruff check app/services/tvscreener_service.py app/services/tvscreener_retry.py app/services/tvscreener_capabilities.py; uv run --group test pytest tests/test_mcp_screen_stocks_tvscreener_contract.py tests/test_tvscreener_crypto.py tests/test_tvscreener_stock_enrichment_fields.py -q (33 passed)\nNot-tested: Full repository test suite
